+++ test/unit/fuzz-tools.test.js
@@ -447,6 +447,7 @@ import * as brouetCryoglobulinemiaV468 from '../../lib/brouet-cryoglobulinemia-v
 import * as steinbrockerRaV469 from '../../lib/steinbrocker-ra-v469.js';
 import * as larsenRaV470 from '../../lib/larsen-ra-v470.js';
 import * as gassMacularHoleV471 from '../../lib/gass-macular-hole-v471.js';
+import * as yerdelPvtV472 from '../../lib/yerdel-pvt-v472.js';
 import * as tbTesting from '../../lib/tb-testing.js';
 
 const MODULES = {
@@ -839,6 +840,7 @@ const MODULES = {
   'steinbrocker-ra-v469.js': steinbrockerRaV469,
   'larsen-ra-v470.js': larsenRaV470,
   'gass-macular-hole-v471.js': gassMacularHoleV471,
+  'yerdel-pvt-v472.js': yerdelPvtV472,
   'tb-testing.js': tbTesting,
 };